LayoutTests/inspector/debugger/source-file.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>

<script>

function test()
{
    var mockSourceMapping = {
        originalToFormatted: function(location)
        {
            var formattedLocation = {};
            formattedLocation.lineNumber = location.lineNumber * 2;
            formattedLocation.columnNumber = location.columnNumber * 2;
            return formattedLocation;
        },

        formattedToOriginal: function(location)
        {
            var originalLocation = {};
            originalLocation.lineNumber = Math.floor(location.lineNumber / 2);
            originalLocation.columnNumber = Math.floor(location.columnNumber / 2);
            return originalLocation;
        }
    };

    var mockScriptFormatter = {
        formatContent: function(mimeType, content, callback)
        {
            callback(content, mockSourceMapping);
        }
    };

    function rl(lineNumber, columnNumber)
    {
        return { lineNumber: lineNumber, columnNumber: columnNumber };
    }

    function checkRawLocation(script, lineNumber, columnNumber, location)
    {
        InspectorTest.assertEquals(script.scriptId, location.scriptId);
        InspectorTest.assertEquals(lineNumber, location.lineNumber);
        InspectorTest.assertEquals(columnNumber, location.columnNumber);
    }

    function checkUILocation(rawSourceCode, lineNumber, columnNumber, location)
    {
        InspectorTest.assertEquals(rawSourceCode.uiSourceCode, location.uiSourceCode);
        InspectorTest.assertEquals(lineNumber, location.lineNumber);
        InspectorTest.assertEquals(columnNumber, location.columnNumber);
    }

    InspectorTest.runTestSuite([
        function testPlainConvertLocation(next)
        {
            var script = new WebInspector.Script("1", "foo.js", 0, 0, 20, 80, undefined, undefined, false);
            var rawSourceCode = new WebInspector.RawSourceCode("id", script);

            checkUILocation(rawSourceCode, 10, 20, rawSourceCode.rawLocationToUILocation(rl(10, 20)));
            checkRawLocation(script, 30, 40, rawSourceCode.uiLocationToRawLocation(30, 40));

            next();
        },

        function testConcatenatedConvertLocation(next)
        {
            var script1 = new WebInspector.Script("1", "foo.js", 10, 20, 30, 40, undefined, undefined, false);
            var script2 = new WebInspector.Script("2", "foo.js", 50, 60, 70, 80, undefined, undefined, false);
            var rawSourceCode = new WebInspector.RawSourceCode("id", script1);
            rawSourceCode.addScript(script2);

            checkUILocation(rawSourceCode, 20, 0, rawSourceCode.rawLocationToUILocation(rl(20, 0)));

            checkRawLocation(script1, 0, 40, rawSourceCode.uiLocationToRawLocation(0, 40));
            checkRawLocation(script1, 20, 0, rawSourceCode.uiLocationToRawLocation(20, 0));
            checkRawLocation(script2, 50, 60, rawSourceCode.uiLocationToRawLocation(50, 60));

            next();
        },

        function testFormattedConvertLocation(next)
        {
            var script = new WebInspector.Script("1", "foo.js", 0, 0, 20, 80, undefined, undefined, false);
            var rawSourceCode = new WebInspector.RawSourceCode("id", script, mockScriptFormatter, true);

            function didCreateSourceMapping()
            {
                checkUILocation(rawSourceCode, 22, 40, rawSourceCode.rawLocationToUILocation(rl(11, 20)));
                checkRawLocation(script, 7, 9, rawSourceCode.uiLocationToRawLocation(14, 19));
                next();
            }
            rawSourceCode.createSourceMappingIfNeeded(didCreateSourceMapping);
        },

        function testConcatenatedFormattedConvertLocation(next)
        {
            var script1 = new WebInspector.Script("1", "foo.js", 10, 20, 30, 40, undefined, undefined, false);
            var script2 = new WebInspector.Script("2", "foo.js", 50, 60, 70, 80, undefined, undefined, false);
            var rawSourceCode = new WebInspector.RawSourceCode("id", script1, mockScriptFormatter, true);
            rawSourceCode.addScript(script2);

            function didCreateSourceMapping()
            {
                checkUILocation(rawSourceCode, 22, 60, rawSourceCode.rawLocationToUILocation(rl(11, 30)));
                checkRawLocation(script1, 12, 10, rawSourceCode.uiLocationToRawLocation(24, 20));
                checkRawLocation(script2, 60, 19, rawSourceCode.uiLocationToRawLocation(121, 38));
                next();
            }
            rawSourceCode.createSourceMappingIfNeeded(didCreateSourceMapping);
        },

        function testUISourceCodeAdded(next)
        {
            var script = new WebInspector.Script("1", "foo.js", 0, 0, 20, 80, undefined, undefined, false);
            var rawSourceCode = new WebInspector.RawSourceCode("id", script);
            function uiSourceCodeAdded(event)
            {
                InspectorTest.assertEquals("foo.js", event.data.url);
            }
            rawSourceCode.addEventListener(WebInspector.RawSourceCode.Events.UISourceCodeAdded, uiSourceCodeAdded, this);

            next();
        }
    ]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests mapping between raw locations and ui locations.</p>

</body>
</html>
